+++ mods/Smithing/mod/data.lua
@@ -3,6 +3,7 @@ require("prototypes.group")
 
 require("prototypes.item.cables")
 
+require("prototypes.anvil")
 
 require("prototypes.recipe.bar_recipe")
 require("prototypes.recipe.intermediate_recipe")
@@ -10,3 +11,5 @@ require("prototypes.recipe.pickaxe_recipe")
 
 require("prototypes.technology.pickaxe_technology")
 require("prototypes.technology.smelting_technology")
+
+require("prototypes.category")