+++ mods/Smithing/mod/prototypes/category.lua
@@ -1,0 +1,9 @@
+data:extend({
+  {
+    type = "recipe-category",
+    name = "rs-smithing"
+  }
+})
+
+table.insert(data.raw["god-controller"].default.crafting_categories, "rs-smithing")
+table.insert(data.raw.character.character.crafting_categories, "rs-smithing")