+++ mods/Smithing/CHANGELOG.md
@@ -2,3 +2,7 @@
 
 ### Changed
 - Moved bars, pickaxes and intermediates from Mining to Smithing
+
+### Added
+- Added anvil entity
+- Added logic to make items "smithable", in a furnace you smelt bars into an unfinished item. On the anvil you then create the actual item